pyproject.toml
[project]
name = "backend-django-hw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "requests>=2.32.4",
    "ruff>=0.12.4",
]
